+++ packages/common/content/ActionDetailMap.ts
