+++ packages/common/content/ActivityDetailMap.ts
