+++ packages/common/content/FacilityDetailMap.ts
