+++ packages/common/content/FacilityTypeDetailMap.ts
