+++ packages/common/content/GameData.ts
@@ -1,13 +1,13 @@
-import { skillDetailMap } from "common/content/skills/SkillDetailMap";
-import { skillExpLevels } from "common/content/skills/SkillExpLevels";
-import { activityDetailMap } from "common/content/activities/ActivityDetailMap";
-import { actionDetailMap } from "common/content/actions/ActionDetailMap";
-import { itemDetailMap } from "common/content/items/ItemDetailMap";
-import { roadDetailMap } from "common/content/worldmap/RoadDetailMap";
-import { npcDetailMap } from "common/content/npcs/NpcDetailMap";
-import { facilityTypeDetailMap } from "common/content/worldmap/FacilityTypeDetailMap";
-import { facilityDetailMap } from "common/content/worldmap/FacilityDetailMap";
-import { locationDetailMap } from "common/content/worldmap/LocationDetailMap";
+import { skillDetailMap } from "common/content/SkillDetailMap";
+import { skillExpLevels } from "common/content/SkillExpLevels";
+import { activityDetailMap } from "common/content/ActivityDetailMap";
+import { actionDetailMap } from "common/content/ActionDetailMap";
+import { itemDetailMap } from "common/content/ItemDetailMap";
+import { roadDetailMap } from "common/content/RoadDetailMap";
+import { npcDetailMap } from "common/content/NpcDetailMap";
+import { facilityTypeDetailMap } from "common/content/FacilityTypeDetailMap";
+import { facilityDetailMap } from "common/content/FacilityDetailMap";
+import { locationDetailMap } from "common/content/LocationDetailMap";
 
 export const gameData = {
   activityDetailMap: activityDetailMap,
