+++ packages/common/content/ItemDetailMap.ts
