+++ packages/common/content/LocationDetailMap.ts
