+++ packages/common/content/NpcDetailMap.ts
