+++ packages/common/content/SkillDetailMap.ts
