+++ packages/common/content/SkillExpLevels.ts
